Export ReadIntoTree
This is to prevent code duplication in the corresponding MerkleTree type
in the Sia crypto package, which reimplimented this functionality in
MerkleTree.ReadSegments. This change will allow MerkleTree.ReadSegments
to wrap ReadIntoTree.

This is being done because the MerkleTree.ReadSegments implementation
had a bug where the last leaf would be padded incorrectly.