implement and test cached trees
Cached trees allow you to build merkle trees from the cached merkle
roots of other trees. The root of a cached tree will be the same as if
you had not done any caching. You can also make storage proofs from
cached trees. The proof will be the same as if you had not done any
cacheing.

This is good for efficiecny when it comes to building storage proofs
over TBs of data. Instead of needing to hash all of the data, you only
need to hash the set of roots you've cached. If the segment size is 64
bytes and the cache size is 4MB, you get a ~100k speedup on hashing.

This is also good for modifying files and being able to quickly and
accurately update the root. If you modify a file, you update all the
affected cached nodes and then you use the cached nodes to compute the
new root using a CachedTree.